requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-cors-middleware==1.3.1
django-templated-mail==1.1.1
djangorestframework==3.8.2
djoser==1.1.5
httplib2==0.11.3
idna==2.6
pytz==2018.4
requests==2.18.4
shortuuid==0.5.0
urllib3==1.22
